endret til ConsumerRecord

+++ Eventhandtering/src/main/resources/schema.sql
@@ -1,9 +1,9 @@
 create table dokumentevent
 (
-    id    int           NOT NULL AUTO_INCREMENT,
+    id    bigint           NOT NULL AUTO_INCREMENT,
     dokumentid  varchar(255)  not null,
     navn  varchar(255)  not null,
-    sideantall numeric(5, 2) not null,
+    sideantall int not null,
     PRIMARY KEY (id),
     UNIQUE (dokumentid)
 );
